pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orion-v3"
version = "3.0.0"
description = "GSS Orion V3 — Adaptive IA Orchestration Engine"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "litellm>=1.40.0",
    "jinja2>=3.1",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.0",
    "ruff>=0.4",
    "httpx",
]

[tool.setuptools.packages.find]
include = ["core*", "brain*", "experts*", "ops*", "portal*"]

[tool.pytest.ini_options]
testpaths = ["ops/tests"]
python_files = "test_*.py"
addopts = "-v --tb=short -n 2"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]
